requirements.txt
numpy
opencv-python
ujson
PySimpleGUI
mediapipe
onnxruntime-gpu==1.13.1